pyproject.toml: relax requirements
Overly strict requirements makes it difficult to resolve packages.
Particularly in this case, many of the projects provide wheels only for
the latest version.  As L2G will be used as a library, it is reasonable
to expect that users may be using slightly different versions of torch,
torch-geometric and other fundamental libraries.

Relaxing numpy requirement will lead to numpy 2 being installed. Most
libraries in the ecosystem support numpy 2 now, so should not be an
issue.